+++ server/.env.example
@@ -1,1 +1,2 @@
 MONGODB_URI = mongodb+srv://<username>:<password>@<cluster>.mongodb.net/?retryWrites=true&w=majority&appName=<appname>
+SECRET = CHANGEME
