+++ server/src/routes/auth.route.js
@@ -1,8 +1,8 @@
-import { Router } from "express"; // Importando o Router do express
-const router = Router(); // Instanciando o Router
+import { Router } from "express";
+const router = Router();
 
 
-import { login } from "../controllers/auth.controller.js"; // Importando o controller de autenticação
-router.post("/", login); //autenticação é feita por meio do POST
+import { login } from "../controllers/auth.controller.js";
+router.post("/", login);
 
-export default router; // Exportando o router para ser utilizado em outros arquivos
+export default router;
